FIR UAST: convert destructuring assignments
