+++ plugins/kotlin/uast/uast-kotlin-fir/testData/legacyRenderLog/DestructuringDeclaration.log.fir.txt
@@ -10,6 +10,20 @@ UFile (package = )
                         UBinaryExpression (operator = <other>)
                             ULiteralExpression (value = "foo")
                             ULiteralExpression (value = 1)
+                    ULocalVariable (name = a)
+                        UAnnotation (fqName = null)
+                        UQualifiedReferenceExpression
+                            USimpleNameReferenceExpression (identifier = var268d41a5)
+                            UCallExpression (kind = UastCallKind(name='method_call'), argCount = 0))
+                                UIdentifier (Identifier (component1))
+                                USimpleNameReferenceExpression (identifier = <anonymous class>, resolvesTo = null)
+                    ULocalVariable (name = b)
+                        UAnnotation (fqName = null)
+                        UQualifiedReferenceExpression
+                            USimpleNameReferenceExpression (identifier = var268d41a5)
+                            UCallExpression (kind = UastCallKind(name='method_call'), argCount = 0))
+                                UIdentifier (Identifier (component2))
+                                USimpleNameReferenceExpression (identifier = <anonymous class>, resolvesTo = null)
                 UDeclarationsExpression
                     ULocalVariable (name = var465978a0)
                         UAnnotation (fqName = null)
@@ -17,3 +31,17 @@ UFile (package = )
                         UBinaryExpressionWithType
                             USimpleNameReferenceExpression (identifier = data)
                             UTypeReferenceExpression (name = kotlin.Pair<? extends java.lang.String,? extends java.lang.String>)
+                    ULocalVariable (name = k)
+                        UAnnotation (fqName = null)
+                        UQualifiedReferenceExpression
+                            USimpleNameReferenceExpression (identifier = var465978a0)
+                            UCallExpression (kind = UastCallKind(name='method_call'), argCount = 0))
+                                UIdentifier (Identifier (component1))
+                                USimpleNameReferenceExpression (identifier = <anonymous class>, resolvesTo = null)
+                    ULocalVariable (name = v)
+                        UAnnotation (fqName = null)
+                        UQualifiedReferenceExpression
+                            USimpleNameReferenceExpression (identifier = var465978a0)
+                            UCallExpression (kind = UastCallKind(name='method_call'), argCount = 0))
+                                UIdentifier (Identifier (component2))
+                                USimpleNameReferenceExpression (identifier = <anonymous class>, resolvesTo = null)
