+++ plugins/kotlin/uast/uast-kotlin-fir/testData/legacyRenderLog/DestructuringDeclaration.render.fir.txt
@@ -1,6 +1,10 @@
 public final class DestructuringDeclarationKt {
     public static final fun foo(@org.jetbrains.annotations.NotNull data: java.lang.Object) : void {
         @null var var268d41a5: kotlin.Pair<? extends java.lang.String,? extends java.lang.Integer> = "foo" <other> 1
+        @null var a: java.lang.String = var268d41a5.<anonymous class>()
+        @null var b: int = var268d41a5.<anonymous class>()
         @null @null var var465978a0: kotlin.Pair<? extends java.lang.String,? extends java.lang.String> = data as kotlin.Pair<? extends java.lang.String,? extends java.lang.String>
+        @null var k: java.lang.String = var465978a0.<anonymous class>()
+        @null var v: java.lang.String = var465978a0.<anonymous class>()
     }
 }
