+++ plugins/kotlin/uast/uast-kotlin-fir/testData/legacyRenderLog/LambdaReturn.render.fir.txt
@@ -32,6 +32,8 @@ public final class LambdaReturnKt {
         })
         var x: kotlin.jvm.functions.Function0<? extends kotlin.Unit> = {
             @org.jetbrains.annotations.NotNull var vardbcd0724: java.util.List<? extends java.lang.Integer> = listOf(1, 2)
+            @org.jetbrains.annotations.NotNull var a: int = vardbcd0724.<anonymous class>()
+            @org.jetbrains.annotations.NotNull var b: int = vardbcd0724.<anonymous class>()
         }
         var y: kotlin.jvm.functions.Function0<? extends kotlin.Unit> = {
             return listOf(1)
