--- plugins/kotlin/uast/uast-kotlin-fir/testData/legacyRenderLog/WhenAndDestructing.log.fir.txt
@@ -1,34 +1,0 @@
-UFile (package = )
-    UClass (name = WhenAndDestructingKt)
-        UMethod (name = getElementsAdditionalResolve)
-            UParameter (name = string)
-                UAnnotation (fqName = org.jetbrains.annotations.NotNull)
-            UBlockExpression
-                UDeclarationsExpression
-                    ULocalVariable (name = arr)
-                        UCallExpression (kind = UastCallKind(name='method_call'), argCount = 2))
-                            UIdentifier (Identifier (listOf))
-                            USimpleNameReferenceExpression (identifier = listOf, resolvesTo = null)
-                            ULiteralExpression (value = "1")
-                            ULiteralExpression (value = "2")
-                USwitchExpression
-                    USimpleNameReferenceExpression (identifier = string)
-                    UExpressionList (when)
-                        USwitchClauseExpressionWithBody
-                            ULiteralExpression (value = "aaaa")
-                            UExpressionList (when_entry)
-                                UYieldExpression
-                                    UReturnExpression
-                                        ULiteralExpression (value = "bindingContext")
-                        USwitchClauseExpressionWithBody
-                            ULiteralExpression (value = "empty-switch")
-                            UExpressionList (when_entry)
-                        USwitchClauseExpressionWithBody
-                            UExpressionList (when_entry)
-                                UDeclarationsExpression
-                                    ULocalVariable (name = var837f2350)
-                                        UAnnotation (fqName = null)
-                                        USimpleNameReferenceExpression (identifier = arr)
-                                UYieldExpression
-                                    UReturnExpression
-                                        USimpleNameReferenceExpression (identifier = bindingContext)
