+++ plugins/kotlin/uast/uast-kotlin-fir/testData/legacyRenderLog/WhenAndDestructing.log.txt
