+++ plugins/kotlin/uast/uast-kotlin-fir/testData/legacyRenderLog/WhenAndDestructing.render.fir.txt
@@ -11,6 +11,8 @@ public final class WhenAndDestructingKt {
 
              -> {
                 @null var var837f2350: java.util.List<? extends java.lang.String> = arr
+                @null var bindingContext: java.lang.String = var837f2350.<anonymous class>()
+                @null var statementFilter: java.lang.String = var837f2350.<anonymous class>()
                 yield return bindingContext
             }
