+++ plugins/kotlin/uast/uast-kotlin-fir/testData/legacyValues/DestructuringDeclaration.values.fir.txt
@@ -4,16 +4,44 @@ UFile (package = ) [public final class DestructuringDeclarationKt {...]
             UParameter (name = data) [@org.jetbrains.annotations.NotNull var data: java.lang.Object]
                 UAnnotation (fqName = org.jetbrains.annotations.NotNull) [@org.jetbrains.annotations.NotNull]
             UBlockExpression [{...}] = Undetermined
-                UDeclarationsExpression [@null var var268d41a5: kotlin.Pair<? extends java.lang.String,? extends java.lang.Integer> = "foo" <other> 1] = Undetermined
+                UDeclarationsExpression [@null var var268d41a5: kotlin.Pair<? extends java.lang.String,? extends java.lang.Integer> = "foo" <other> 1...@null var b: int = var268d41a5.<anonymous class>()] = Undetermined
                     ULocalVariable (name = var268d41a5) [@null var var268d41a5: kotlin.Pair<? extends java.lang.String,? extends java.lang.Integer> = "foo" <other> 1]
                         UAnnotation (fqName = null) [@null]
                         UBinaryExpression (operator = <other>) ["foo" <other> 1] = Undetermined
                             ULiteralExpression (value = "foo") ["foo"] = "foo"
                             ULiteralExpression (value = 1) [1] = (long)1
-                UDeclarationsExpression [@null @null var var465978a0: kotlin.Pair<? extends java.lang.String,? extends java.lang.String> = data as kotlin.Pair<? extends java.lang.String,? extends java.lang.String>] = Undetermined
+                    ULocalVariable (name = a) [@null var a: java.lang.String = var268d41a5.<anonymous class>()]
+                        UAnnotation (fqName = null) [@null]
+                        UQualifiedReferenceExpression [var268d41a5.<anonymous class>()] = external <anonymous class>()()
+                            USimpleNameReferenceExpression (identifier = var268d41a5) [var268d41a5] = external var268d41a5()
+                            UCallExpression (kind = UastCallKind(name='method_call'), argCount = 0)) [<anonymous class>()] = external <anonymous class>()()
+                                UIdentifier (Identifier (component1)) [UIdentifier (Identifier (component1))]
+                                USimpleNameReferenceExpression (identifier = <anonymous class>, resolvesTo = null) [<anonymous class>] = external <anonymous class>()()
+                    ULocalVariable (name = b) [@null var b: int = var268d41a5.<anonymous class>()]
+                        UAnnotation (fqName = null) [@null]
+                        UQualifiedReferenceExpression [var268d41a5.<anonymous class>()] = external <anonymous class>()()
+                            USimpleNameReferenceExpression (identifier = var268d41a5) [var268d41a5] = external var268d41a5()
+                            UCallExpression (kind = UastCallKind(name='method_call'), argCount = 0)) [<anonymous class>()] = external <anonymous class>()()
+                                UIdentifier (Identifier (component2)) [UIdentifier (Identifier (component2))]
+                                USimpleNameReferenceExpression (identifier = <anonymous class>, resolvesTo = null) [<anonymous class>] = external <anonymous class>()()
+                UDeclarationsExpression [@null @null var var465978a0: kotlin.Pair<? extends java.lang.String,? extends java.lang.String> = data as kotlin.Pair<? extends java.lang.String,? extends java.lang.String>...@null var v: java.lang.String = var465978a0.<anonymous class>()] = Undetermined
                     ULocalVariable (name = var465978a0) [@null @null var var465978a0: kotlin.Pair<? extends java.lang.String,? extends java.lang.String> = data as kotlin.Pair<? extends java.lang.String,? extends java.lang.String>]
                         UAnnotation (fqName = null) [@null]
                         UAnnotation (fqName = null) [@null]
                         UBinaryExpressionWithType [data as kotlin.Pair<? extends java.lang.String,? extends java.lang.String>] = Undetermined
                             USimpleNameReferenceExpression (identifier = data) [data] = Undetermined
                             UTypeReferenceExpression (name = kotlin.Pair<? extends java.lang.String,? extends java.lang.String>) [kotlin.Pair<? extends java.lang.String,? extends java.lang.String>] = Undetermined
+                    ULocalVariable (name = k) [@null var k: java.lang.String = var465978a0.<anonymous class>()]
+                        UAnnotation (fqName = null) [@null]
+                        UQualifiedReferenceExpression [var465978a0.<anonymous class>()] = external <anonymous class>()()
+                            USimpleNameReferenceExpression (identifier = var465978a0) [var465978a0] = external var465978a0()
+                            UCallExpression (kind = UastCallKind(name='method_call'), argCount = 0)) [<anonymous class>()] = external <anonymous class>()()
+                                UIdentifier (Identifier (component1)) [UIdentifier (Identifier (component1))]
+                                USimpleNameReferenceExpression (identifier = <anonymous class>, resolvesTo = null) [<anonymous class>] = external <anonymous class>()()
+                    ULocalVariable (name = v) [@null var v: java.lang.String = var465978a0.<anonymous class>()]
+                        UAnnotation (fqName = null) [@null]
+                        UQualifiedReferenceExpression [var465978a0.<anonymous class>()] = external <anonymous class>()()
+                            USimpleNameReferenceExpression (identifier = var465978a0) [var465978a0] = external var465978a0()
+                            UCallExpression (kind = UastCallKind(name='method_call'), argCount = 0)) [<anonymous class>()] = external <anonymous class>()()
+                                UIdentifier (Identifier (component2)) [UIdentifier (Identifier (component2))]
+                                USimpleNameReferenceExpression (identifier = <anonymous class>, resolvesTo = null) [<anonymous class>] = external <anonymous class>()()
