+++ plugins/kotlin/uast/uast-kotlin/src/org/jetbrains/uast/kotlin/KotlinConverter.kt
@@ -148,18 +148,4 @@ internal object KotlinConverter : BaseKotlinConverter {
             else -> expr<UExpression>(build(::UnknownKotlinExpression))
         }}
     }
-
-    internal fun KtPsiFactory.createAnalyzableExpression(text: String, context: PsiElement): KtExpression =
-        createAnalyzableProperty("val x = $text", context).initializer ?: error("Failed to create expression from text: '$text'")
-
-    internal fun KtPsiFactory.createAnalyzableProperty(text: String, context: PsiElement): KtProperty =
-        createAnalyzableDeclaration(text, context)
-
-    internal fun <TDeclaration : KtDeclaration> KtPsiFactory.createAnalyzableDeclaration(text: String, context: PsiElement): TDeclaration {
-        val file = createAnalyzableFile("dummy.kt", text, context)
-        val declarations = file.declarations
-        assert(declarations.size == 1) { "${declarations.size} declarations in $text" }
-        @Suppress("UNCHECKED_CAST")
-        return declarations.first() as TDeclaration
-    }
 }
